+++ openapi/constants.js
@@ -1,0 +1,52 @@
+const partialUpdateList = new Set([
+  'User',
+  'UserProfile'
+]);
+
+const propertyDetailsList = [
+  { path: 'FactorDevice.links', skip: true, skipReason: 'Not currently supported' },
+  { path: 'Link.hints', skip: true, skipReason: 'Not currently supported' },
+  { path: 'User._links', skip: true, skipReason: 'Not currently supported' },
+  { path: 'UserGroup._embedded', skip: true, skipReason: 'Not currently supported' },
+  { path: 'UserGroup._links', skip: true, skipReason: 'Not currently supported' },
+  { path: 'UserGroupStats._links', skip: true, skipReason: 'Not currently supported' },
+  
+  { path: 'ActivationToken.activationToken', rename: 'token', renameReason: '.NET type name and member name cannot be identical' },
+  { path: 'TempPassword.tempPassword', rename: 'password', renameReason: '.NET type name and member name cannot be identical' },
+
+  { path: 'CallFactor.profile', hidesBaseMember: true },
+  { path: 'EmailFactor.profile', hidesBaseMember: true },
+  { path: 'HardwareFactor.profile', hidesBaseMember: true },
+  { path: 'PushFactor.profile', hidesBaseMember: true },
+  { path: 'SecurityQuestionFactor.profile', hidesBaseMember: true },
+  { path: 'SmsFactor.profile', hidesBaseMember: true },
+  { path: 'TokenFactor.profile', hidesBaseMember: true },
+  { path: 'TotpFactor.profile', hidesBaseMember: true },
+  { path: 'WebFactor.profile', hidesBaseMember: true },
+];
+
+const operationSkipList = [
+  { id: 'forgotPassword', reason: 'Revisit in alpha2 (#62)'},
+  { id: 'addRoleToUser', reason: 'Revisit in alpha2 (#102)'},
+];
+
+const modelMethodSkipList = [
+  { path: 'User.changePassword', reason: 'Implemented as ChangePasswordAsync(options)' },
+  { path: 'User.changeRecoveryQuestion', reason: 'Implemented as ChangeRecoveryQuestionAsync(options)'},
+  { path: 'User.forgotPassword', reason: 'Revisit in alpha2 (#64)'},
+  { path: 'User.addRole', reason: 'Implemented as a custom method'},
+  { path: 'User.listAppLinks', reason: 'Implemented as IUser.AppLinks' },
+  { path: 'User.listRoles', reason: 'Implemented as IUser.Roles' },
+  { path: 'User.listGroups', reason: 'Implemented as IUser.Groups' },
+  { path: 'User.removeRole', reason: 'Add back in alpha2 (#64)' },
+  { path: 'User.listGroupTargetsForRole', reason: 'Too complex for IUser, leave on IUserClient' },
+  { path: 'User.addGroupTargetToRole', reason: 'Too complex for IUser, leave on IUserClient' },
+  { path: 'User.removeGroupTargetFromRole', reason: 'Too complex for IUser, leave on IUserClient' },
+  { path: 'User.resetPassword', reason: 'Simplified as IUser.ResetPasswordAsync(bool)' },
+  { path: 'Group.listUsers', reason: 'Implemented as IGroup.Users' },
+];
+
+module.exports.partialUpdateList = partialUpdateList;
+module.exports.propertyDetailsList = propertyDetailsList;
+module.exports.operationSkipList = operationSkipList;
+module.exports.modelMethodSkipList = modelMethodSkipList;
